requirements.txt
python-telegram-bot==13.7
--force-reinstall
--no-cache-dir